+++ lib/my_files/init.dart
@@ -1,8 +1,22 @@
 library init_my_files;
 
+export 'dart:async';
+export 'dart:io';
+export 'package:flutter/services.dart';
+export 'package:open_document/open_document.dart';
+export 'package:share_plus/share_plus.dart';
+
+export 'components/alert_customized_for_decisions.dart';
+export 'components/button_share_confirm_cancel.dart';
+export 'components/header_my_folder_file.dart';
 export 'components/my_files_core.dart';
 export 'components/my_files_dialog.dart';
 export 'components/my_files_items.dart';
+export 'components/slidable_my_file_item.dart';
 export 'model/custom_file_system_entity.dart';
 export 'model/style_my_file.dart';
+export 'model/extract_zip.dart';
 export 'my_files_screen.dart';
+export 'controller_my_files.dart';
+export '../open_document_exception.dart';
+